+++ backend/src/main/java/com/campito/backend/dto/TransaccionDTOResponse.java
@@ -1,5 +1,6 @@
 package com.campito.backend.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.UUID;
@@ -9,7 +10,7 @@ import com.campito.backend.model.TipoTransaccion;
 public record TransaccionDTOResponse(
     Long id,
     LocalDate fecha,
-    Float monto,
+    BigDecimal monto,
     TipoTransaccion tipo,
     String descripcion,
     String nombreCompletoAuditoria,
